+++ Code/net/PhasedBroadcast.h
@@ -153,7 +153,7 @@ namespace hemelb
          * Returns the number of the iteration, as an integer between inclusive-0 and
          * exclusive-2 * (the tree depth)
          */
-        unsigned int Get0IndexedIterationNumber() const;
+        unsigned long Get0IndexedIterationNumber() const;
 
         /**
          * Note that depths are 0-indexed. I.e. a tree with a single node has
